Add additional build script to Kernel project
This script copies the XNU binary into the root directory, and names it “mach_kernel”. This change was prompted by a change in the 10.10 XNU build process that installs the kernel into a different location with a different filename. This change will break the booter, so I use this script to restore the old semantics.